requirements.txt
python-ldap>=2.4.25
flask>=0.11.1
